Add args-test
